### 2021-04-26 #442 @makinin
rdbhapi package: An R package to simplify API access to data on higher education in Norway
Submitting Author: Marija Ninic (@makinin)
Other Package Authors: Sigve Hofland Berge (@sigve-berge-hofland)
Repository:  https://github.com/makinin/rdbhapi

---

```
Package: rdbhapi
Title: Interface to 'DBH-API'
Version: 2.1.0
Authors@R: c(person(given = "Marija",family = "Ninic",role = c("cre", "aut"),email = "marija.ninic@nsd.no", comment=c(ORCID ="0000-0001-8484-5817")),person(given=c("Sigve", "Hofland"), family="Berge", role = "aut"))
Description: Downloads 'DBH-API'<https://dbh.nsd.uib.no/tjenester.action> data as dataframes. These data are in machine-readable format and provides information about education, finances, research and publications.
    httr,
    rjson,
    stats
```

- Please indicate which category or categories from our [package fit policies](https://ropensci.github.io/dev_guide/policies.html#package-categories) this package falls under: (Please check an appropriate box below.:

	- [ x ] data retrieval

- Explain how and why the package falls under these categories (briefly, 1-2 sentences).
This package falls under the category of data retrieval because it simplify API access from to data on higher education in Norway, as provided by NSD - Norwegian Centre for Research Data. This package read data from API into R dataframe.

   The target audience are students and researchers who needs data about education in their work.

 No there is not.

- No
