### 2021-05-03 #442 @makinin
How to add Technical checks, Publication options, MEE Options, Code of conduct, can not find them in template
