### 2021-05-03 #442 @maelle
@makinin I mean that in the body of the template there are questions like "Explain how and why the package falls under these categories (briefly, 1-2 sentences)." and no answer yet. :slightly_smiling_face:

Also, were you intending to fill a pre-submission inquiry rather than a submission? The template is https://github.com/ropensci/software-review/blob/master/.github/ISSUE_TEMPLATE/B-submit-a-presubmission-inquiry.md (and also has some of the questions you haven't answered yet).
