### 2021-05-05 #442 @noamross
@makinin Thanks for making those changes and submitting a pre-submission inquirty. This does fit into our package Aims and Scope. When you want to start the review process, please fill out the full submission template here: https://github.com/ropensci/software-review/issues/new?assignees=&labels=&template=A-submit-software-for-review.md
